Modif input shape

+++ GAN2.py
@@ -59,6 +59,7 @@ class GAN():
         self.d_optimizer=Adam(self.learning_rate,self.beta1)
         self.g_optimizer=Adam(self.learning_rate*self.fact_g_lr,self.beta1)
         self.build_model()
+
     def build_model(self):
 
         # We use the discriminator
@@ -187,12 +188,12 @@ class GAN():
         for epoch in range(0, self.epoch):
             print("starting epoch {}".format(epoch))
             for idx in range(start_batch_id, self.num_batches):
-                d_noise_real = np.float32(random.uniform(self.real_label_smoothing[0],self.real_label_smoothing[1]))  # Add noise on the loss
-                d_noise_fake = np.float32(random.uniform(self.fake_label_smoothing[0],self.fake_label_smoothing[1]))  # Add noise on the loss
+                d_noise_real = random.uniform(self.real_label_smoothing[0],self.real_label_smoothing[1])  # Add noise on the loss
+                d_noise_fake = random.uniform(self.fake_label_smoothing[0],self.fake_label_smoothing[1]) # Add noise on the loss
                 #print(idx * self.batch_size, (idx + 1) * self.batch_size)
-                batch_input = self.data_X[idx * self.batch_size:(idx + 1) * self.batch_size]  # the input
+                batch_input = self.data_X[idx * self.batch_size:(idx + 1) * self.batch_size].astype(np.float32)  # the input
                 # print("batch_input ite {} shape {} ".format(idx,batch_input.shape))
-                batch_gt = self.data_y[idx * self.batch_size:(idx + 1) * self.batch_size]  # the Ground Truth images
+                batch_gt = self.data_y[idx * self.batch_size:(idx + 1) * self.batch_size].astype(np.float32)  # the Ground Truth images
                 # Generate a batch of new images
                 gen_imgs = self.generator.predict(batch_input) #.astype(np.float32)
                 print("gen_img",gen_imgs.dtype)
